requirements.txt
gradio
sentence-transformers
faiss-cpu
openai